Web:Client: split router to client, portal settings and confirm routes
