+++ packages/client/src/routes/index.js
@@ -1,0 +1,7 @@
+import PortalSettingsRouter from "./portalSettings";
+import ClientRouter from "./client";
+import ConfirmRouter from "./confirm";
+
+const routes = [...ClientRouter, PortalSettingsRouter, ...ConfirmRouter];
+
+export default routes;
