+++ packages/client/src/routes/utils.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+export const NotFoundError = () => {
+  return (
+    <PrivateRoute>
+      <ErrorBoundary>
+        <Error404 />
+      </ErrorBoundary>
+    </PrivateRoute>
+  );
+};
